+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to Audit.NET and its extensions will be documented in this f
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [12.0.0] - 2018-??
+### Added
+- Audit.Core: Added a global audit switch off `Configuration.AuditDisabled`.
+- Audit.Core: Added `NullDataProvider` as an anternative to disable the audit logging.
+
+
 ## [11.2.0] - 2018-04-11
 ### Changed
 - Audit.NET.MongoDB: Fix #114 - MongoDB Dataprovider Date serialization. Changing serialization mechanism to store the .NET DateTime as mongo datetime.